+++ racing_data/entity.py
@@ -36,11 +36,6 @@ class Entity(dict):
             self.property_cache = {}
 
         if key not in self.property_cache:
-            self.property_cache = source_method(*source_args, **source_kwargs)
+            self.property_cache[key] = source_method(*source_args, **source_kwargs)
 
         return self.property_cache[key]
-
-    def is_equivalent_to(self, other_entity):
-        """This entity is equivalent to other_entity if both have been sourced from the same URL"""
-
-        return self['url'] == other_entity['url']